add key256 in upgrade keys
